Private Routes and Playlist beginnings
Private routes seem to be working. Trying to navigate to a protected route take the user back to the default page with the login button. 

Playlist rendering seems to be working at this point for spotify.Successfully retrieves data from the api and displays the results in material-ui's <List> component.